Remove stateRef conversion to Long.

+++ components/uniqueness/backing-store-impl/src/main/kotlin/net/corda/uniqueness/backingstore/impl/JPABackingStoreImpl.kt
@@ -191,7 +191,7 @@ open class JPABackingStoreImpl @Activate constructor(
                 )
                     .setParameter("txAlgo", txId.algorithm)
                     .setParameter("txId", txId.bytes)
-                    .setParameter("stateIndex", stateIndex.toLong())
+                    .setParameter("stateIndex", stateIndex)
                     .resultList as List<UniquenessStateDetailEntity>
 
                 existing.firstOrNull()?.let { stateEntity ->
@@ -280,7 +280,7 @@ open class JPABackingStoreImpl @Activate constructor(
                         UniquenessStateDetailEntity(
                             stateRef.txHash.algorithm,
                             stateRef.txHash.bytes,
-                            stateRef.stateIndex.toLong(),
+                            stateRef.stateIndex,
                             null, // Unconsumed
                             null // Unconsumed
                         )
@@ -300,7 +300,7 @@ open class JPABackingStoreImpl @Activate constructor(
                         .setParameter("consumingTxId", consumingTxId.bytes)
                         .setParameter("issueTxAlgo", stateRef.txHash.algorithm)
                         .setParameter("issueTxId", stateRef.txHash.bytes)
-                        .setParameter("stateIndex", stateRef.stateIndex.toLong())
+                        .setParameter("stateIndex", stateRef.stateIndex)
 
                     val updatedRowCount = safeUpdate.executeUpdate()
